Add pseudo-break method to pass failing break spec
Currently named :please_to_go_away_so_fast, it throws out
to a catch in the iterator method. There's no way to take
care of this in :bastardize/:debastardize, is there? I
doubt it, which means a catch wrap will have to be included
in every iterator method. Pretty lame, eh?