Fix up `#undefine_methods` and `#index`
`#undefine_methods` was not actually doing anything.
It was just rescuing every attempt to undefine a
nonexistent method called `meth`. Eventually I'll
need to have it raise a proper error---right now
it will print the name of a method it failed to
undefine.
Also, I eliminated an unnecessary `#bastardize` call
in `#index`.