Remove specs for JRuby/Rubinius-specific methods
JRuby's Array includes an instance method :iter_for_each.
Rubinius's Array includes :total and :total=.

All 3 of these methods are now skipped in the specs where an
IterableArray is tested to ensure it responds to all of Array's instance
methods, as I don't know enough about what they do or the need for them
to implement them in IterableArray.